+++ selfdrive/ui/qt/maps/map.h
@@ -4,8 +4,6 @@
 
 #include <QGeoCoordinate>
 #include <QGestureEvent>
-#include <QHash>
-#include <QHBoxLayout>
 #include <QLabel>
 #include <QMap>
 #include <QMapboxGL>
@@ -15,7 +13,6 @@
 #include <QPushButton>
 #include <QScopedPointer>
 #include <QString>
-#include <QTextDocument>
 #include <QVBoxLayout>
 #include <QWheelEvent>
 
@@ -23,42 +20,8 @@
 #include "common/params.h"
 #include "common/util.h"
 #include "selfdrive/ui/ui.h"
-
-class MapInstructions : public QWidget {
-  Q_OBJECT
-
-private:
-  QLabel *distance;
-  QLabel *primary;
-  QLabel *secondary;
-  QLabel *icon_01;
-  QHBoxLayout *lane_layout;
-  bool is_rhd = false;
-  std::vector<QLabel *> lane_labels;
-  QHash<QString, QPixmap> pixmap_cache;
-
-public:
-  MapInstructions(QWidget * parent=nullptr);
-  void buildPixmapCache();
-  QString getDistance(float d);
-  void updateInstructions(cereal::NavInstruction::Reader instruction);
-};
-
-class MapETA : public QWidget {
-  Q_OBJECT
-
-public:
-  MapETA(QWidget * parent=nullptr);
-  void updateETA(float seconds, float seconds_typical, float distance);
-
-private:
-  void paintEvent(QPaintEvent *event) override;
-  void showEvent(QShowEvent *event) override { format_24h = param.getBool("NavSettingTime24h"); }
-
-  bool format_24h = false;
-  QTextDocument eta_doc;
-  Params param;
-};
+#include "selfdrive/ui/qt/maps/map_eta.h"
+#include "selfdrive/ui/qt/maps/map_instructions.h"
 
 class MapWindow : public QOpenGLWidget {
   Q_OBJECT
